# src/main/resources/swaggerDoc.yaml
swagger: "2.0"
info:
  description: "Documento referente ao rest do sistema Ponto Security"
  version: "1.0.0"
  title: "Ponto Security rest"
host: "audax.mobi"
basePath: "/rest/"
tags:
- name: "Teste"
  description: "Description"
schemes:
- "http"
paths:
  /login:
    post:
      tags:
      - "Teste"
      summary: "Loggar com o funcionario"
      description: "Entrar no sistema"
      operationId: "Testelogin"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Aparelhos bluetooth e senha do funcionario"
        required: true
        schema:
          $ref: "#/definitions/ReflectionHelper2"
      responses:
        "403":
          description: "desc"
          schema:
            $ref: "#/definitions/ErroMensage"
  /receberDadosRecipientes:
    post:
      tags:
      - "Teste"
      summary: "Loggar com o funcionario"
      description: "Entrar no sistema"
      operationId: "TestereceberDadosRecipientes"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Aparelhos bluetooth e senha do funcionario"
        required: true
        schema:
          $ref: "#/definitions/ReflectionHelper"
      responses:
        "200":
          description: "desc"
  /listRecipientes:
    get:
      tags:
      - "Teste"
      summary: "Loggar com o funcionario"
      description: "Entrar no sistema"
      operationId: "listRecipientes"
      produces:
        - "application/json"
      parameters: []
      responses:
        "400":
          description: "desc"
  /listDadosRecipiente/{id}:
    get:
      tags:
      - "Teste"
      summary: "Loggar com o funcionario"
      description: "Entrar no sistema"
      operationId: "listDadosRecipiente/{id}"
      produces:
        - "application/json"
      parameters:
      - in: path
        name: id
        required: true
        type: integer
      responses:
        "400":
          description: "desc"
        "200":
          description: "desc"
          schema:
            $ref: "#/definitions/ReflectionHelper4"
definitions:
  ReflectionHelper:
    type: "object"
    required:
    - "age"
    properties:
      age:
        type: "integer"
        example: 42
      name:
        type: "string"
        example: "Texto de exemplo"
      deptName:
        type: "string"
        example: "Texto de exemplo"
      empID:
        type: "integer"
        example: 42
      teste:
        type: "array"
        items: 
          $ref: "#/definitions/ReflectionHelper2"
  ReflectionHelper2:
    type: "object"
    required:
    - "age"
    properties:
      age:
        type: "integer"
        example: 42
      name:
        type: "string"
        example: "Texto de exemplo"
      deptName:
        type: "string"
        example: "Texto de exemplo"
      empID:
        type: "integer"
        example: 42
      teste:
        type: "array"
        items: 
          $ref: "#/definitions/ReflectionHelper3"
  ReflectionHelper3:
    type: "object"
    required:
    - "age"
    properties:
      age:
        type: "integer"
        example: 42
      name:
        type: "string"
        example: "Texto de exemplo"
      deptName:
        type: "string"
        example: "Texto de exemplo"
      empID:
        type: "integer"
        example: 42
  ErroMensage:
    type: "object"
    properties:
      code:
        type: "integer"
        example: 42
      mensage:
        type: "string"
        example: "Texto de exemplo"
  ReflectionHelper4:
    type: "object"
    required:
    - "age"
    properties:
      age:
        type: "integer"
        example: 42
      name:
        type: "string"
        example: "Texto de exemplo"
      deptName:
        type: "string"
        example: "Texto de exemplo"
      empID:
        type: "integer"
        example: 42